Feature #1 completed and closed, refactored various part of the code
