+++ public/index.html
@@ -28,6 +28,7 @@
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
+    <div id="modal-root"></div>
     <div id="root"></div>
     <!--
       This HTML file is a template.
