+++ src/App.css
@@ -3,11 +3,11 @@
 :root{
   --color1: #bae8e8;
   --color2: #7fe8e8;
+  --color3: goldenrod;
   --border-radius: 10px;
 }
 
 .container {
-  /* border: 1px solid green; */
   max-width: 756px;
   margin: auto;
   padding: .8em;
@@ -24,8 +24,26 @@ header h1{
   color: white;
   font-family: 'Press Start 2P', cursive;
   font-size: 1.5em;
-  /* font-family: 'Bebas Neue', cursive; */
+}
 
+.create-invoice-container{
+  text-align: center;
+}
+
+.create-invoice-button{
+  padding: 1em;
+  background-color: transparent;
+  font-size: 1em;
+  font-weight: bolder;
+  color: var(--color3);
+  border: 1px solid var(--color3);
+  border-radius: var(--border-radius);
+  transition: 0.3s
+}
+
+.create-invoice-button:hover{
+  background-color: var(--color3);
+  color: white;
 }
 
 /* Main Section styling */
@@ -62,38 +80,3 @@ input[type="submit"] {
   font-size: 1em;
   background-color: var(--color1);
 }
-
-/* Table Styling */
-
-.table{
-  margin-top: 3em;
-  overflow: auto;
-}
-
-.table-header, .table-record{
-  min-width: 470px;
-  display: flex;
-  justify-content: space-between;
-}
-
-.table-header div, .table-record div{
-  /* border: 1px solid palevioletred; */
-  text-align: center;
-  flex: 1 0 0;
-  padding: 0.5em 0.4em;
-  border: 1px solid goldenrod;
-}
-
-.table-header{
-  background-color: goldenrod;
-  color: white;
-}
-
-.table-record{
-  background-color: lemonchiffon;
-}
-
-.table-header div{
-  font-weight: bold;
-}
-
