+++ src/App.js
@@ -1,6 +1,8 @@
 import Table from './Table';
 import './App.css';
 import { useState } from 'react';
+import Modal from './Modal';
+import Form from './Form';
 
 const data = [
   {
@@ -30,58 +32,16 @@ const data = [
 ]
 
 function App() {
-  const [inputData, setInputData] = useState({
-    invoiceId: Math.floor(Math.random() * 100000),
-    invoiceDate: "",
-    unitPrice: "",
-    quantity: "",
-    totalAmount: "",
-    status: "Pending"
-  });
-  console.log(inputData)
-  const formatDate = (date) => {
-    return new Date(date).toISOString().split('T')[0];
+  // State variables
+  
+  const [isOpen, setIsOpen] = useState(false);
+
+  const getData = (formData) => {
+    data.push(formData)
   }
 
-
-  const handleOnChange = (field, value) => {
-    switch (field) {
-      case "invoiceDate":
-        setInputData((prevState) => {
-          return { ...prevState, invoiceDate: formatDate(value) }
-        });
-        break;
-      case "unitPrice":
-        setInputData((prevState) => {
-          return { ...prevState, unitPrice: value, totalAmount: prevState.quantity * value }
-        });
-        break;
-      case "quantity":
-        setInputData((prevState) => {
-          return { ...prevState, quantity: value, totalAmount: prevState.unitPrice * value }
-        });
-        break;
-      case "status":
-        setInputData((prevState) => {
-          return { ...prevState, status: value }
-        });
-        break;
-      default:
-        break;
-    }
-  }
-
-  const handleOnSubmit = (e) => {
-    e.preventDefault();
-    data.push(inputData);
-    setInputData({
-      invoiceId: Math.floor(Math.random() * 1000),
-      invoiceDate: "",
-      unitPrice: "",
-      quantity: "",
-      totalAmount: "",
-      status: "Pending"
-    })
+  const handleClose = () => {
+    setIsOpen(!isOpen);
   }
 
   return (
@@ -90,64 +50,15 @@ function App() {
         <h1>Invoicer</h1>
       </header>
       <main className="container app">
-        <form className='invoice-form' onSubmit={handleOnSubmit}>
-          <div className='input-field'>
-            <label htmlFor="invoice-date">Invoice Date</label>
-            <input
-              onChange={(e) => handleOnChange("invoiceDate", e.target.value)}
-              value={inputData.invoiceDate}
-              id='invoice-date'
-              type="date"
-              required
-            />
-          </div>
-          <div className='input-field'>
-            <label htmlFor="unit-price">Unit Price</label>
-            <input
-              onChange={(e) => handleOnChange("unitPrice", e.target.value)}
-              value={inputData.unitPrice}
-              id='unit-price'
-              type="number"
-              min={0}
-              required
-            />
-          </div>
-          <div className='input-field'>
-            <label htmlFor="quantity">Quantity</label>
-            <input
-              onChange={(e) => handleOnChange("quantity", e.target.value)}
-              value={inputData.quantity}
-              id='quantity'
-              type="number" min={1} max={100}
-              required
-            />
-          </div>
-          <div className='input-field'>
-            <label htmlFor="total-amount">Total amount</label>
-            <input
-              readOnly
-              value={inputData.totalAmount}
-              id='total-amount'
-              type="number"
-            />
-          </div>
-          <div className='input-field'>
-            <label htmlFor="status">Status</label>
-            <select
-              onChange={(e) => handleOnChange("status", e.target.value)}
-              value={inputData.status}
-              id="status"
-              name="status"
-            >
-              <option value="Pending">Pending</option>
-              <option value="Paid">Paid</option>
-            </select>
-          </div>
-          <div className='input-field'>
-            <input type="submit" />
-          </div>
-        </form>
-
+        <div className='create-invoice-container'>
+          <button className='create-invoice-button' onClick={handleClose}>Create New Invoice</button>
+        </div>
+        {isOpen
+          &&
+          <Modal closeModal={handleClose}>
+            <Form closeForm={handleClose} sendData={getData} />
+          </Modal>
+        }
         <Table data={data} />
       </main>
     </>
