+++ src/Modal.module.css
@@ -1,0 +1,29 @@
+.overlay{
+    background: rgba(142, 142, 142, 0.5);
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    min-height: 100vh;
+    padding: 0.8em;
+}
+
+.modal{
+    background-color: white;
+    position: relative;
+    max-width: 450px;
+    margin: auto;
+    margin-top: 100px;
+    padding: 1em;
+    border-radius: 10px;
+    animation: animate-modal 0.4s ease-in;
+}
+
+@keyframes animate-modal {
+    0%{
+        margin-top: -600px;
+    }
+    100%{
+        margin-top: 100px;
+    }
+}
