+++ src/Table.js
@@ -1,10 +1,11 @@
 import Record from "./Record";
+import styles from "./Table.module.css";
 
 const Table = (props) => {
     const data = props.data;
     return (
-        <div className="table">
-            <div className='table-header'>
+        <div className={styles.table}>
+            <div className={styles['table-header']}>
                 <div>Invoice Number</div>
                 <div>Invoice Date</div>
                 <div>Unit price</div>
@@ -13,7 +14,7 @@ const Table = (props) => {
                 <div>Status</div>
             </div>
             {data.map((record) => {
-                return <Record key={Math.random()} record={record} />
+                return <Record key={record.invoiceId} record={record} />
             })}
         </div>
     )
